# api/openapi.yaml
openapi: 3.1.0
info:
  title: Solitaire Matchmaking API
  version: 0.7.0
  description: |
    Implemented HTTP endpoints and the outbound game-backend event contract.
    Every HTTP response includes `X-Request-ID`, `Cache-Control: no-store` and
    `X-Content-Type-Options: nosniff` headers.
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
paths:
...
  /v1/capabilities:
    get:
      tags: [Integration]
      summary: Read implemented service capabilities
      operationId: getCapabilities
      security:
        - apiBearerAuth: []
      responses:
        "200":
          description: Current service capabilities.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"
              example:
                service: solitaire-matchmaking
                stage: release_candidate
                rating_enabled: true
                matchmaking_enabled: true
                ticket_lifecycle_enabled: true
                result_ingestion_enabled: true
                outbox_delivery_enabled: true
                metrics_enabled: true
                planned_room_sizes: [5, 6, 7]
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/rooms/{room_id}:
    parameters:
      - name: room_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
    get:
      tags: [Integration]
      summary: Read room composition and lifecycle state
      operationId: getRoom
      security:
        - apiBearerAuth: []
      responses:
        "200":
          description: Current room metadata, lifecycle timestamps and seat-ordered members.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Room does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: {error: room_not_found}
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    apiBearerAuth:
      type: http
      scheme: bearer
      description: API token configured as `API_TOKEN`.
    outboxBearerAuth:
      type: http
      scheme: bearer
      description: Delivery token configured as `OUTBOX_DELIVERY_TOKEN`.
  headers:
    RequestID:
      description: Opaque request correlation identity generated by the service.
      schema:
        type: string
        minLength: 1
  responses:
    Unauthorized:
      description: Bearer authentication failed.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            const: Bearer
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: {error: unauthorized}
    InternalError:
      description: Operation failed unexpectedly.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: {error: internal_error}
  schemas:
...
    Capabilities:
      type: object
      additionalProperties: false
      required: [service, stage, rating_enabled, matchmaking_enabled, ticket_lifecycle_enabled, result_ingestion_enabled, outbox_delivery_enabled, metrics_enabled, planned_room_sizes]
      properties:
        service:
          type: string
          const: solitaire-matchmaking
        stage:
          type: string
          const: release_candidate
        rating_enabled:
          type: boolean
          const: true
        matchmaking_enabled:
          type: boolean
          const: true
        ticket_lifecycle_enabled:
          type: boolean
          const: true
        result_ingestion_enabled:
          type: boolean
          const: true
        outbox_delivery_enabled:
          type: boolean
          const: true
        metrics_enabled:
          type: boolean
          const: true
        planned_room_sizes:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: integer
            enum: [5, 6, 7]
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          enum: [unauthorized, invalid_ticket, invalid_cancellation, invalid_rating_query, tournament_not_found, ticket_not_found, rating_not_found, idempotency_conflict, ticket_not_queued, invalid_result, room_not_found, result_conflict, room_not_collecting, result_deadline_passed, participants_mismatch, internal_error]
...
    RoomState:
      type: object
      additionalProperties: false
      required: [room_id, tournament_id, tournament_version, mode_id, policy_version, rating_model_version, scoring_rules_version, settlement_version, deck_id, capacity, status, aggregate_version, created_at, fill_deadline, members]
      properties:
        room_id: {type: string, minLength: 1}
        tournament_id: {type: string, minLength: 1}
        tournament_version: {type: string, minLength: 1}
        mode_id: {type: string, minLength: 1}
        policy_version: {type: string, minLength: 1}
        rating_model_version: {type: string, minLength: 1}
        scoring_rules_version: {type: string, minLength: 1}
        settlement_version: {type: string, minLength: 1}
        deck_id: {type: string, minLength: 1}
        capacity:
          type: integer
          enum: [5, 6, 7]
        status:
          type: string
          enum: [forming, collecting, completed, expired, cancelled]
        aggregate_version:
          type: integer
          format: int64
          minimum: 1
        created_at: {type: string, format: date-time}
        fill_deadline: {type: string, format: date-time}
        filled_at: {type: string, format: date-time}
        result_deadline: {type: string, format: date-time}
        completed_at: {type: string, format: date-time}
        expired_at: {type: string, format: date-time}
        cancelled_at: {type: string, format: date-time}
        members:
          type: array
          maxItems: 7
          description: Current room members ordered by seat.
          items:
            $ref: "#/components/schemas/RoomMember"
    RoomMember:
      type: object
      additionalProperties: false
      required: [ticket_id, player_id, session_id, seat, status, assigned_at]
      properties:
        ticket_id: {type: string, minLength: 1}
        player_id: {type: string, minLength: 1}
        session_id: {type: string, minLength: 1}
        seat: {type: integer, minimum: 1, maximum: 7}
        status:
          type: string
          enum: [allocated, playing, submitted, forfeited]
        assigned_at: {type: string, format: date-time}
        started_at: {type: string, format: date-time}
        submitted_at: {type: string, format: date-time}
        forfeited_at: {type: string, format: date-time}
...